% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voipred.R
\name{calc_NB_moments}
\alias{calc_NB_moments}
\title{Calculates the first two moments of the bivariate distribution of NB_model and NB_all}
\usage{
calc_NB_moments(Y, pi, z, weights = NULL)
}
\arguments{
\item{Y}{Vector of the binary response variable}

\item{pi}{Vector of predicted risks}

\item{z}{Decision threshold at which the NBs are calculated}

\item{weights}{Optinal - observation weights}
}
\value{
Two means, two SDs, and one correlation coefficient. First element is for the model and second is for treating all
}
\description{
Calculates the first two moments of the bivariate distribution of NB_model and NB_all
}
