% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb.R
\name{Cb.cox}
\alias{Cb.cox}
\title{Cb calculations for a Cox proportional hazard model.}
\usage{
Cb.cox(reg_object, tx_var, semi_parametric = FALSE, time = 1)
}
\arguments{
\item{reg_object}{An object of class 'coxph' that contains the model. IMPORTANT: the coxph function call for fitting the model must have model=TRUE as an input argument such that the underlying data becomes part of the returned object (this is the case by default for glm)}

\item{tx_var}{A string containing the name of the treatment indicator variable.}

\item{semi_parametric}{Optional (default=FALSE). If TRUE, the semi-parametric estimator for Cb will be returned.}

\item{time}{Optional (default=1). The value of time at which Cb is calculated.}
}
\value{
This function returns an object of class Cb_output, which includes Cb as a member.
}
\description{
Cb calculations for a Cox proportional hazard model.
}
\examples{
data("rct_data")
# Create an event indicator and update the tte (time-to-event) variable to be 
# equal to follow-up time for censored individuals.
event<-(!is.na(rct_data[,'tte']))*1
ids<-which(event==0)
rct_data[ids,'tte']<-rct_data[ids,'time']
rct_data['event']<-event

reg.coxph<-coxph(Surv(time=tte,event=event) ~ tx + tx:female + tx:age + sgrq + prev_hosp + prev_ster + fev1, data=rct_data, model=TRUE)
res.coxph<-Cb.coxph(reg.coxph,tx_var = "tx",semi_parametric = T)
}
