% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb.R
\name{Cb.poisson}
\alias{Cb.poisson}
\title{Cb calculations for a Poisson (or negative binomial) regression model.}
\usage{
Cb.poisson(reg_object, tx_var, semi_parametric = FALSE, time = 1)
}
\arguments{
\item{reg_object}{An object of class 'glm' that contains the resuls of the Poisson/NegBin model.}

\item{tx_var}{A string containing the name of the treatment indicator variable.}

\item{semi_parametric}{Optional (default=FALSE). If TRUE, the semi-parametric estimator for Cb will be returned.}

\item{time}{Optional (default=1). The value of time at which Cb is calculated.}
}
\value{
This function returns an object of class Cb_output, which includes Cb as a member.
}
\description{
Cb calculations for a Poisson (or negative binomial) regression model.
}
\examples{
data("rct_data")
reg<-glm(formula = n_exac ~ tx + sgrq + prev_hosp + prev_ster + fev1, 
         data = rct_data, 
         family = poisson(link="log"), 
         offset=ln_time)
res.Poisson<-Cb.Poisson(reg,tx_var = "tx", semi_parametric = T)
res.Poisson
}
