% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb.R
\name{Cb.simple}
\alias{Cb.simple}
\title{Simple, parametric calculation of Cb.}
\usage{
Cb.simple(B)
}
\arguments{
\item{B}{A numberic vector.}
}
\value{
This function returns Cb as 1-E(B1)/E(max(B2,B3)), where B1, B2, and B3 are random draws from the empirical distribution of B.
}
\description{
Simple, parametric calculation of Cb.
}
\examples{
b <- runif(1000)
Cb.simple(b)
}
