% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_data.R
\docType{data}
\name{dev_data}
\alias{dev_data}
\title{model development data}
\format{
A data frame with 500 rows and 5 variables:
\itemize{
  \item{age}{age}
  \item{severity}{dooshvari}
  \item{sex}{dooshvari}
  \item{comorbidity}{dooshvari}
  \item{y}{dooshvari}
}
}
\source{
Simulated
}
\description{
A dataset containing sample model development data
}
\keyword{datasets}
