% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gusto.R
\docType{data}
\name{gusto}
\alias{gusto}
\title{Anonymized data from the gusto trial}
\format{
A data frame with 40830 rows and 29 variables:
\itemize{
  \item{day30}{whether death happened by day 30 after intervention}
  \item{sho}{whether cardiac shock was present}
  \item{hig}{whether the patient hat high blood pressure}
  \item{dia}{whether the patient had diabetes}
  \item{hrt}{whether the patient was on hormone replacement therapies}
}
}
\source{
Internet
}
\description{
A dataset containing anonymized data from the gusto trial
}
\keyword{datasets}
