% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mROC.R
\name{mAUC}
\alias{mAUC}
\title{Takes in a mROC object and calculates the area under the curve}
\usage{
mAUC(mROC_obj)
}
\arguments{
\item{mROC_obj}{An object of class mROC}
}
\value{
Returns the area under the mROC curve
}
\description{
Takes in a mROC object and calculates the area under the curve
}
