% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mROC}
\alias{mROC}
\title{Takes in a vector of probabilities and returns mROC values (Tp,TP, in an object of class mROC)}
\usage{
mROC(p, ordered = F)
}
\arguments{
\item{p}{A numberic vector of probabilities.}

\item{ordered}{Optional, if the vector p is ordered from small to large (if not the function will do it; TRUE is to facilitate fast computations).}
}
\value{
This function returns an object of class mROC.
}
\description{
Takes in a vector of probabilities and returns mROC values (Tp,TP, in an object of class mROC)
}
