% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mROC_analysis}
\alias{mROC_analysis}
\title{Main eRoc analysis: draws the ROC and eROC. inference=0: no inference, inference=1: p-value, inference=2: p-value and 95%CI}
\usage{
mROC_analysis(y, p, inference = 0, n_sim, fast = TRUE)
}
\description{
Main eRoc analysis: draws the ROC and eROC. inference=0: no inference, inference=1: p-value, inference=2: p-value and 95%CI
}
