% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_max_trunc_bvn.R
\name{mu_max_trunc_bvn}
\alias{mu_max_trunc_bvn}
\title{Calculates the expected value of the maximum of two random variables with zero-truncated bivariate normal distirbution
Takes a vector of mean and a 2X2 covariance matrix}
\usage{
mu_max_trunc_bvn(mu, sigma)
}
\arguments{
\item{mu}{A vector of means (of length 2)}

\item{sigma}{A 2X2 covariance matrix.}
}
\value{
A scalar value for the expected value
}
\description{
Calculates the expected value of the maximum of two random variables with zero-truncated bivariate normal distirbution
Takes a vector of mean and a 2X2 covariance matrix
}
