% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interceptAdj.R
\name{odds_adjust}
\alias{odds_adjust}
\title{Title Update a prediction model for a binary outcome by multiplying a fixed odd-ratio to the predicted odds.}
\usage{
odds_adjust(p0, p1, v)
}
\arguments{
\item{p0}{Mean of observed risk or predicted risk in development sample.}

\item{p1}{Mean of observed risk in target population.}

\item{v}{Variance of predicted risk in development sample.}
}
\value{
Returns a correction factor that can be applied to the predicted 
odds in order to update the predictions for a new target population.
}
\description{
Title Update a prediction model for a binary outcome by multiplying a fixed odd-ratio to the predicted odds.
}
