% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interceptAdj.R
\name{pred_summary_stat}
\alias{pred_summary_stat}
\title{Title Estimate mean and variance of prediction based on model calibration output.}
\usage{
pred_summary_stat(calibVector)
}
\arguments{
\item{calibVector}{Vector of predicted probability of risk per decile or percentile (e.g., from a calibration plot).}
}
\value{
Returns mean and variance of predictions based on the predicted probabilities.
}
\description{
Title Estimate mean and variance of prediction based on model calibration output.
}
