% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_data.R
\docType{data}
\name{rct_data}
\alias{rct_data}
\title{simulated randomized clinical trial data}
\format{
A data frame with 1108 rows and 10 variables:
\itemize{
  \item{tx}{whether received the treatment}
  \item{female}{whether the patient is female or not}
  \item{age}{the age of the patient}
  \item{prev_hosp}{dooshvari}
  \item{prev_ster}{dooshvari}
  \item{fev1}{forced expiratory volume in one second}
  \item{sgrq}{St. George's Questionnaire Score, a measure of symptoms}
  \item{time}{dooshvari}
  \item{tte}{time to event}
  \item{n_exac}{number of exacerbations}
  
}
}
\source{
Simulated
}
\description{
A dataset containing sample RCT data
}
\keyword{datasets}
