% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calc_mROC_stats}
\alias{calc_mROC_stats}
\title{Calculates the absolute surface between the empirical and expected ROCs}
\usage{
calc_mROC_stats(y, p, ordered = F, fast = T)
}
\arguments{
\item{y}{y vector of binary responses}

\item{p}{p vector of predicted probabilities (same length as y)}

\item{ordered}{defaults to false}

\item{fast}{defaults to true}
}
\description{
Calculates the absolute surface between the empirical and expected ROCs
}
