% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calc_mROC_stats}
\alias{calc_mROC_stats}
\title{Calculates the asolute surface between the empirical and expected ROCs}
\usage{
calc_mROC_stats(y, p, ordered = F, fast = T)
}
\description{
Calculates the asolute surface between the empirical and expected ROCs
}
