% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_data.R
\docType{data}
\name{dev_data}
\alias{dev_data}
\title{model development data}
\format{
A data frame with 500 rows and 5 variables:
\itemize{
  \item{age}{age}
  \item{severity}{whether or not the disease was severe}
  \item{sex}{binary sex variable, 1 for female and 0 for male}
  \item{comorbidity}{whether or not comorbidities are present}
  \item{y}{response variable}
}
}
\source{
Simulated
}
\description{
A dataset containing sample model development data
}
\keyword{datasets}
