% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_data.R
\docType{data}
\name{dev_data}
\alias{dev_data}
\title{model development data}
\format{
A data frame with 500 rows and 5 variables:
\describe{
  \item{age}{price, in US dollars}
  \item{severity}{weight of the diamond, in carats}
  \item{sex}{weight of the diamond, in carats}
  \item{comorbidity}{weight of the diamond, in carats}
  \item{y}{weight of the diamond, in carats}
}
}
\source{
Simulated
}
\usage{
dev_data
}
\description{
A dataset containing sample model development data
}
\keyword{datasets}
