% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mROC_analysis}
\alias{mROC_analysis}
\title{Main eRoc analysis that plots ROC and eROC}
\usage{
mROC_analysis(y, p, inference = 0, n_sim, fast = TRUE)
}
\arguments{
\item{y}{y dooshvari}

\item{p}{p dooshvari}

\item{inference}{0 for no inference, 1 for p-value only, and 2 for p-value and 95 percent CI.}

\item{n_sim}{number of simulations}

\item{fast}{defaults to true}
}
\value{
plot and analysis dooshvari
}
\description{
Main eRoc analysis that plots ROC and eROC
}
