% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mROC_inference}
\alias{mROC_inference}
\title{Statistical inference for comparing empirical and expectec ROCs. If CI=TRUE then also returns pointwise CIs}
\usage{
mROC_inference(
  y,
  p,
  n_sim = 1e+05,
  CI = FALSE,
  aux = FALSE,
  fast = TRUE,
  conditional = FALSE
)
}
\arguments{
\item{y}{vector of binary response values}

\item{p}{vector of probabilities}

\item{n_sim}{number of Monte Carlo simulations to calculate p-value}

\item{CI}{whether confidence interval should be alculated for each point of mROC}
}
\value{
Returns the aurea under the mROC curve
}
\description{
Statistical inference for comparing empirical and expectec ROCs. If CI=TRUE then also returns pointwise CIs
}
