% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_data.R
\docType{data}
\name{val_data}
\alias{val_data}
\title{model validation data}
\format{
A data frame with 400 rows and 5 variables:
\itemize{
  \item{age}{age of the patient}
  \item{severity}{dooshvari}
  \item{sex}{dooshvari}
  \item{comorbidity}{dooshvari}
  \item{y}{dooshvari}
}
}
\source{
Simulated
}
\description{
A dataset containing sample model validation data
}
\keyword{datasets}
